requirements.txt
Flask
requests
joblib
pandas
uvicorn
fastapi
scikit-learn==1.3.2
numpy
scipy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client